requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
Django==2.1.3
djangorestframework==3.9.0
idna==2.7
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.0
pytz==2018.7
requests==2.20.1
six==1.11.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
